+++ src/model/model.ts
@@ -41,6 +41,6 @@ export * from './import';
 export * from './inline-document';
 export * from './literal';
 export * from './property';
-export * from './package';
+export {Package, QueryOptions as PackageQueryOptions} from './package';
 export * from './resolvable';
 export * from './source-range';